fix(bioemu-benchmarks): fix multiconf results class (#20)
Multiconf results specification introduced `apo` / `holo`
differentiation for crypticpocket benchmark. This broke some of the
plotting and results aggregation routines (or one version was just
ignored).

Updated `MulticonfResults` to iterate over both cases when the
crypticpocket benchmark is encountered and updated labeling of results
and plots.

(The changes made it through the unit tests because we were not testing
for this particular benchmark in the CLI tests.)

---------

Co-authored-by: Michael Gastegger <mgastegger@microsoft.com>