feat(bioemu-benchmarks): make filtering faster (#23)
Add prefiltering based on Ca distances before checking contacts on all
atoms which was the most time intensive step in the procedure. Filter
behavior remains the same, but speed should be a factor of 10 faster and
memory requirement lower (especially for larger systems).

#### Example on domain motion
Old:

![image](https://github.com/user-attachments/assets/c8c01d59-2328-4ad2-bcc7-25996a9fe8be)
New:

![image](https://github.com/user-attachments/assets/e5c0e2bc-e29c-4a31-9f85-f09246b3d387)

>[!NOTE]
> This filtering assumes that all atoms in a residue will be distorted
together with the Ca. It will e.g. not detect a single N from a residue
colliding with something far away if the remaining atoms behave
normally. However, this situation should not happen as we use a rigid
frame based representation and MD data should be fine as well due to the
harmonic potentials.

---------

Co-authored-by: Michael Gastegger <mgastegger@microsoft.com>